refactor(FR-1492): extract SharedMemoryFormItems from ResourceAllocationFormItems (#4303)
Resolves #4302 ([FR-1492](https://lablup.atlassian.net/browse/FR-1492))

## Summary
This PR refactors the ResourceAllocationFormItems component structure to improve code organization and maintainability.

## Changes
- **Moved** `ResourceAllocationFormItems` component from `react/src/components/` to `react/src/components/SessionFormItems/` directory for better organization
- **Extracted** shared memory form logic into a separate `SharedMemoryFormItems` component (~240 lines) to improve code modularity and reusability
- **Updated** import paths in all dependent files and test files
- **Reduced** main component complexity by ~335 lines while maintaining all existing functionality

## Benefits
- Better code organization with grouped session form components
- Improved separation of concerns between resource allocation and shared memory management
- More maintainable and testable code structure
- Easier to extend and modify individual form sections independently

## Testing
- All existing tests pass
- Linting and formatting checks passed
- No functional changes - purely a refactoring

[FR-1492]: https://lablup.atlassian.net/browse/FR-1492?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
