+++ react/src/components/ServiceLauncherPageContent.tsx
@@ -26,12 +26,12 @@ import ImageEnvironmentSelectFormItems, {
   ImageEnvironmentFormInput,
 } from './ImageEnvironmentSelectFormItems';
 import InputNumberWithSlider from './InputNumberWithSlider';
+import ResourceNumber from './ResourceNumber';
 import ResourceAllocationFormItems, {
   AUTOMATIC_DEFAULT_SHMEM,
   RESOURCE_ALLOCATION_INITIAL_FORM_VALUES,
   ResourceAllocationFormValue,
-} from './ResourceAllocationFormItems';
-import ResourceNumber from './ResourceNumber';
+} from './SessionFormItems/ResourceAllocationFormItems';
 import VFolderLazyView from './VFolderLazyView';
 import VFolderSelect from './VFolderSelect';
 import VFolderTableFormItem from './VFolderTableFormItem';
