+++ react/src/components/SessionFormItems/ResourceAllocationFormItems.tsx
@@ -1,45 +1,32 @@
-import { ResourceAllocationFormItemsQuery } from '../__generated__/ResourceAllocationFormItemsQuery.graphql';
+import { ResourceAllocationFormItemsQuery } from '../../__generated__/ResourceAllocationFormItemsQuery.graphql';
 import {
   addNumberWithUnits,
   compareNumberWithUnits,
   convertToBinaryUnit,
-} from '../helper';
-import { useSuspendedBackendaiClient, useUpdatableState } from '../hooks';
-import { useResourceSlotsDetails } from '../hooks/backendai';
-import { useCurrentKeyPairResourcePolicyLazyLoadQuery } from '../hooks/hooksUsingRelay';
-import { useCurrentProjectValue } from '../hooks/useCurrentProject';
-import { useEventNotStable } from '../hooks/useEventNotStable';
+} from '../../helper';
+import { useSuspendedBackendaiClient, useUpdatableState } from '../../hooks';
+import { useResourceSlotsDetails } from '../../hooks/backendai';
+import { useCurrentKeyPairResourcePolicyLazyLoadQuery } from '../../hooks/hooksUsingRelay';
+import { useCurrentProjectValue } from '../../hooks/useCurrentProject';
+import { useEventNotStable } from '../../hooks/useEventNotStable';
 import {
   MergedResourceLimits,
   ResourcePreset,
   useResourceLimitAndRemaining,
-} from '../hooks/useResourceLimitAndRemaining';
-import AgentSelect from './AgentSelect';
-import BAISelect from './BAISelect';
-import DynamicUnitInputNumber from './DynamicUnitInputNumber';
-import DynamicUnitInputNumberWithSlider from './DynamicUnitInputNumberWithSlider';
+} from '../../hooks/useResourceLimitAndRemaining';
+import AgentSelect from '../AgentSelect';
+import BAISelect from '../BAISelect';
+import DynamicUnitInputNumberWithSlider from '../DynamicUnitInputNumberWithSlider';
 import {
   Image,
   ImageEnvironmentFormInput,
-} from './ImageEnvironmentSelectFormItems';
-import InputNumberWithSlider from './InputNumberWithSlider';
-import QuestionIconWithTooltip from './QuestionIconWithTooltip';
-import ResourceGroupSelect from './ResourceGroupSelect';
-import ResourcePresetSelect from './ResourcePresetSelect';
+} from '../ImageEnvironmentSelectFormItems';
+import InputNumberWithSlider from '../InputNumberWithSlider';
+import ResourceGroupSelect from '../ResourceGroupSelect';
+import ResourcePresetSelect from '../ResourcePresetSelect';
+import SharedMemoryFormItems from './SharedMemoryFormItems';
 import { CaretDownOutlined, ReloadOutlined } from '@ant-design/icons';
-import {
-  Button,
-  Card,
-  Col,
-  ConfigProvider,
-  Divider,
-  Form,
-  Radio,
-  Row,
-  Slider,
-  Switch,
-  theme,
-} from 'antd';
+import { Button, Card, Col, Divider, Form, Radio, Row, theme } from 'antd';
 import { BAIFlex } from 'backend.ai-ui';
 import _ from 'lodash';
 import React, { Suspense, useEffect, useMemo, useTransition } from 'react';
@@ -83,7 +70,7 @@ export interface ResourceAllocationFormValue {
   agent?: string;
 }
 
-type MergedResourceAllocationFormValue = ResourceAllocationFormValue &
+export type MergedResourceAllocationFormValue = ResourceAllocationFormValue &
   ImageEnvironmentFormInput;
 
 interface ResourceAllocationFormItemsProps {
@@ -850,286 +837,24 @@ const ResourceAllocationFormItems: React.FC<
                         );
                       }}
                     </Form.Item>
-                    <Form.Item
-                      noStyle
-                      dependencies={[
-                        ['resource', 'mem'],
-                        ['resource', 'shmem'],
-                        ['enabledAutomaticShmem'],
-                      ]}
-                    >
-                      {({ getFieldValue }) => {
-                        const mem = getFieldValue(['resource', 'mem']) || '0g';
-                        const shmem =
-                          getFieldValue(['resource', 'shmem']) || '0g';
-                        const memUnitResult = convertToBinaryUnit(
-                          mem,
-                          'auto',
-                          2,
-                        );
-                        const shmemUnitResult = convertToBinaryUnit(
-                          shmem,
-                          'auto',
-                          2,
-                        );
-                        const appMemUnitResult = convertToBinaryUnit(
-                          _.max([
-                            0,
-                            (convertToBinaryUnit(mem, 'm')?.number || 0) -
-                              (convertToBinaryUnit(shmem, 'm')?.number || 0),
-                          ]) + 'm',
-                          memUnitResult?.unit || '',
-                        );
-
-                        return (
-                          <BAIFlex direction="column" align="stretch" gap="xs">
-                            <BAIFlex direction="row" gap={'sm'}>
-                              <ConfigProvider
-                                theme={{
-                                  components: {
-                                    Slider: {
-                                      railBg: token.colorWarningBorderHover,
-                                      railHoverBg:
-                                        token.colorWarningBorderHover,
-                                      trackBg: token.colorSuccessBorderHover,
-
-                                      trackHoverBg:
-                                        token.colorSuccessBorderHover,
-                                      railSize: token.fontSize,
-                                    },
-                                  },
-                                }}
-                              >
-                                <Slider
-                                  style={{
-                                    flex: 1,
-                                    margin: 0,
-                                    cursor: 'default',
-                                    padding: 0,
-                                  }}
-                                  styles={{
-                                    handle: {
-                                      display: 'none',
-                                      top: 2,
-                                    },
-                                    root: {
-                                      height: '1em',
-                                    },
-                                  }}
-                                  step={0.001}
-                                  value={appMemUnitResult?.number}
-                                  // Set to 1 to fix UI update issue where slider doesn't rerender when both value and max are 0
-                                  max={memUnitResult?.number || 1}
-                                />
-                              </ConfigProvider>
-                            </BAIFlex>
-                            <BAIFlex
-                              direction="row"
-                              gap={'xxs'}
-                              justify="between"
-                              wrap="wrap"
-                              style={{
-                                minHeight: token.controlHeightSM,
-                              }}
-                            >
-                              <BAIFlex gap={'xxs'}>
-                                <div
-                                  style={{
-                                    height: token.fontSize,
-                                    width: token.fontSize,
-                                    backgroundColor:
-                                      token.colorSuccessBorderHover,
-                                  }}
-                                ></div>
-                                {t('session.launcher.ApplicationMemory', {
-                                  value: appMemUnitResult?.value,
-                                })}
-                              </BAIFlex>
-                              <BAIFlex gap={'xxs'}>
-                                <div
-                                  style={{
-                                    height: token.fontSize,
-                                    width: token.fontSize,
-                                    backgroundColor:
-                                      token.colorWarningBorderHover,
-                                  }}
-                                ></div>
-                                {getFieldValue('enabledAutomaticShmem') &&
-                                  t('session.launcher.SharedMemory', {
-                                    value: shmemUnitResult?.value,
-                                  })}
-                                <Form.Item
-                                  noStyle
-                                  name={['resource', 'shmem']}
-                                  // initialValue={'0g'}
-                                  // label={t('session.launcher.SharedMemory')}
-                                  hidden={form.getFieldValue(
-                                    'enabledAutomaticShmem',
-                                  )}
-                                  dependencies={[['resource', 'mem']]}
-                                  rules={[
-                                    {
-                                      required: true,
-                                      message: t('general.ValueRequired', {
-                                        name: t(
-                                          'session.launcher.SharedMemory',
-                                        ),
-                                      }),
-                                    },
-                                    {
-                                      warningOnly: true,
-                                      validator: async (
-                                        rule,
-                                        value: string,
-                                      ) => {
-                                        const applicationMem =
-                                          appMemUnitResult?.value;
-                                        const shmem = value;
-
-                                        if (
-                                          _.isEmpty(applicationMem) ||
-                                          _.isEmpty(shmem)
-                                        ) {
-                                          return Promise.resolve();
-                                        }
-
-                                        if (
-                                          (convertToBinaryUnit(
-                                            applicationMem,
-                                            'm',
-                                          )?.number || 0) <
-                                          (convertToBinaryUnit(shmem, 'm')
-                                            ?.number || 0) *
-                                            2
-                                        ) {
-                                          throw t(
-                                            'session.launcher.SHMEMShouldBeLessThanHalfOfAppMemory',
-                                          );
-                                        } else {
-                                          return Promise.resolve();
-                                        }
-                                      },
-                                    },
-                                    {
-                                      validator: async (
-                                        rule,
-                                        value: string,
-                                      ) => {
-                                        if (
-                                          _.isEmpty(
-                                            getFieldValue('resource')?.mem,
-                                          ) ||
-                                          _.isEmpty(value) ||
-                                          compareNumberWithUnits(
-                                            getFieldValue('resource')?.mem,
-                                            value,
-                                          ) >= 0
-                                        ) {
-                                          return Promise.resolve();
-                                        } else {
-                                          throw t(
-                                            'resourcePreset.SHMEMShouldBeSmallerThanMemory',
-                                          );
-                                        }
-                                      },
-                                    },
-                                  ]}
-                                >
-                                  <DynamicUnitInputNumber
-                                    // shmem max is mem max
-                                    // min={resourceLimits.shmem?.min}
-                                    min={resourceLimits.shmem?.min}
-                                    size="small"
-                                    // max={resourceLimits.mem?.max || '0g'}
-                                    addonBefore={'SHM'}
-                                    max={
-                                      form.getFieldValue(['resource', 'mem']) ||
-                                      '0g'
-                                    }
-                                    style={{
-                                      width: 200,
-                                    }}
-                                    onChange={() => {
-                                      if (
-                                        baiClient._config
-                                          .allowCustomResourceAllocation
-                                      ) {
-                                        form.setFieldValue(
-                                          'allocationPreset',
-                                          'custom',
-                                        );
-                                      }
-                                    }}
-                                  />
-                                </Form.Item>
-                                <BAIFlex direction="row" gap="xs">
-                                  <Form.Item
-                                    noStyle
-                                    name={'enabledAutomaticShmem'}
-                                    valuePropName="checked"
-                                  >
-                                    <Switch
-                                      size="small"
-                                      title="auto"
-                                      checkedChildren={t('general.Auto')}
-                                      unCheckedChildren={t('general.Manual')}
-                                      onChange={(checked) => {
-                                        if (checked) {
-                                          runShmemAutomationRule(
-                                            form.getFieldValue([
-                                              'resource',
-                                              'mem',
-                                            ]) || '0g',
-                                          );
-                                        }
-                                        if (
-                                          baiClient._config
-                                            .allowCustomResourceAllocation
-                                        ) {
-                                          form.setFieldValue(
-                                            'allocationPreset',
-                                            'custom',
-                                          );
-                                        }
-                                      }}
-                                    />
-                                  </Form.Item>
-                                  <QuestionIconWithTooltip
-                                    title={
-                                      <BAIFlex direction="column">
-                                        {t(
-                                          'session.launcher.AutoSharedMemoryTooltip',
-                                        )}
-                                        <Divider
-                                          style={{
-                                            margin: 0,
-                                            marginBlock: token.marginSM,
-                                            backgroundColor:
-                                              token.colorBorderSecondary,
-                                          }}
-                                        />
-                                        <Trans
-                                          i18nKey={
-                                            'session.launcher.DescSharedMemory'
-                                          }
-                                        />
-                                        <br />
-                                        <br />
-                                        <Trans
-                                          i18nKey={
-                                            'session.launcher.DescSharedMemoryContext'
-                                          }
-                                        />
-                                      </BAIFlex>
-                                    }
-                                  />
-                                </BAIFlex>
-                              </BAIFlex>
-                            </BAIFlex>
-                          </BAIFlex>
-                        );
+                    <SharedMemoryFormItems
+                      min={resourceLimits.shmem?.min}
+                      onChangeResourceShmem={() => {
+                        if (baiClient._config.allowCustomResourceAllocation) {
+                          form.setFieldValue('allocationPreset', 'custom');
+                        }
                       }}
-                    </Form.Item>
+                      onChangeAutomaticShmem={(checked) => {
+                        if (checked) {
+                          runShmemAutomationRule(
+                            form.getFieldValue(['resource', 'mem']) || '0g',
+                          );
+                        }
+                        if (baiClient._config.allowCustomResourceAllocation) {
+                          form.setFieldValue('allocationPreset', 'custom');
+                        }
+                      }}
+                    />
                   </Form.Item>
                 )}
